# Provide zeroconf option to bind to only the default interface